requirements.txt
# =============================================================================
# PERSONAL AI ASSISTANT - DEPENDENCIES (Windows Compatible)
# =============================================================================

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1

# AI & OpenAI Integration
openai==1.45.0
tenacity==8.2.3

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP & CORS
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Templating
jinja2==3.1.2

# Data Processing & Analysis
pandas==2.1.3
numpy==1.24.3

# Date & Time Utilities
python-dateutil==2.8.2

# Environment & Configuration
typing-extensions==4.8.0